Add Aqua and Restrict to Julia ^1.2
Restriction to Julia ^1.2 because Aqua found that the test dependencies
aren't recorded in 1.0/1.1 format. Since I was unable to get the test
env working on 1.0 even with the right specification and a passing Aqua
test, I decided to restrict compatability to the versions with the new
test env handling. Considering that versions before 1.6 are no longer
supported, this seems acceptable.